backend/requirements.txt
flask-bcrypt==1.0.1
pyjwt==2.3.0
flask-cors==4.0.0